+++ app/src/main/java/com/example/newsapp2/api/Destinations.kt
@@ -1,0 +1,8 @@
+package com.example.newsapp2.api
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+object CategoriesScreen
+@Serializable
+class NewsScreen2(val categoryApiId: String)
