+++ app/src/main/java/com/example/newsapp2/ui/theme/Color.kt
@@ -13,3 +13,4 @@ val Pink40 = Color(0xFF7D5260)
 val black=Color(0xFF171717)
 val white=Color(0xFFFFFFFF)
 val gray=Color(0xFFA0A0A0)
+val blackWithOpacity50=Color(0x80000000)
