+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,16 @@
     <string name="news_app_signature">news app signature</string>
     <string name="search_icon">search icon</string>
     <string name="by">By</string>
+    <string name="good_morning_here_is_some_news_for_you">Good Morning\nHere is Some News For You</string>
+    <string name="category">category</string>
+    <string name="general">General</string>
+    <string name="business">Business</string>
+    <string name="sports">Sports</string>
+    <string name="technology">Technology</string>
+    <string name="entertainment">Entertainment</string>
+    <string name="health">Health</string>
+    <string name="science">Science</string>
+    <string name="news_category">news category</string>
+    <string name="view_all">View all</string>
+    <string name="category_view_all">category view all</string>
 </resources>
